+++ nilearn/regions/__init__.py
@@ -2,7 +2,8 @@
 The :mod:`nilearn.regions` class module includes region extraction
 procedure on a 4D statistical/atlas maps and its function.
 """
-from .region_extractor import connected_regions, RegionExtractor
+from .region_extractor import (connected_regions, RegionExtractor,
+                               extract_regions_labels_img)
 from .signal_extraction import (
     img_to_signals_labels, signals_to_img_labels,
     img_to_signals_maps, signals_to_img_maps,
@@ -10,6 +11,7 @@ from .signal_extraction import (
 
 __all__ = [
     'connected_regions', 'RegionExtractor',
+    'extract_regions_labels_img',
     'img_to_signals_labels', 'signals_to_img_labels',
     'img_to_signals_maps', 'signals_to_img_maps',
 ]