+++ doc/whats_new.rst
@@ -5,6 +5,21 @@
 now 0.15.
 
 **Dropped support for Python 2.6**
+=======
+X.X.X
+=====
+
+Changelog
+---------
+
+Enhancements
+............
+
+    - A new function :func:`nilearn.regions.extract_regions_labels_img` can be
+      used for extraction of biggest connected components represented as
+      same label to regions apart with each region labelled as unique label.
+      This function is integrated and can be found in regions module in
+      nilearn.
 
 
 Changelog